Allow for listing buckets with any key.
This change allows for the probe-bucket-sign URL, which is used when
adding a host to see if a key is valid.

Another change uses that in the ListBuckets API. I'd like to find a better
way, so that only buckets that match the bucket key are seen, but so far
I'm unable to think of a way to do that without boiling the
checkRequestAuthType ocean.