Fix ListBuckets to filter matching keys
Listing buckets should return any and only buckets that match
the given API key. The tricky part about this was mostly piping
the parsed access key out to the ListBucketsHandler. Rather than
changing the signature of some functions called in a lot of
places, I created wrapper functions with the old names and signatures,
and then called the inner function with additional return values in
the couple places where I needed them.